agents/iafactory-operator/requirements.txt
# IA Factory Operator - Video Editing Agent
# Production-ready requirements

# === API Framework ===
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-multipart==0.0.9

# === Database ===
sqlmodel==0.0.22
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.10

# === Queue / Background Tasks ===
redis==5.2.0
rq==1.16.2
celery==5.4.0

# === Video Processing ===
ffmpeg-python==0.2.0
moviepy==1.0.3
opencv-python==4.10.0.84
scenedetect==0.6.4
pillow==11.0.0

# === AI / LLM ===
anthropic==0.40.0
openai==1.55.0
openai-whisper==20240930

# === Storage ===
boto3==1.35.0
aiofiles==24.1.0

# === Utilities ===
python-dotenv==1.0.1
httpx==0.27.2
structlog==24.4.0
tenacity==9.0.0

# === Testing ===
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.2
